use size_t rather than unsigned int for dsi and asp cmd/datalen

+++ include/atalk/dsi.h
@@ -66,7 +66,7 @@ typedef struct DSI {
   u_int16_t serverID, clientID;
   u_int8_t *status, commands[DSI_CMDSIZ], data[DSI_DATASIZ];
   int statuslen;
-  unsigned int datalen, cmdlen;
+  size_t datalen, cmdlen;
   size_t read_count, write_count;
   int asleep; /* client won't reply AFP 0x7a ? */
   /* inited = initialized?, child = a child?, noreply = send reply? */
